# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmartBudget API 설계 명세",
    "version": "1.3.0",
    "description": "인증·서버 버전·가계부·소비 분석·OCR에서 구현해야 할 API 계약의 단일 기준인 OpenAPI 3.1 문서입니다. 2026-09-24 현재 Python 서버는 공통 code 응답, 서버 버전 조회와 계정 PATCH·DELETE를 포함한 인증 API를 구현했습니다. 월별 예산 분리, 거래·리포트·OCR API는 아직 구현되지 않았습니다. 구현이 완료될 때까지 실행 중 서버의 /docs와 /openapi.json은 현재 구현 상태를 보여주고, 이 파일은 목표 계약을 나타냅니다.\n\n날짜·일시 형식 규칙: 달력 날짜는 ISO 8601 YYYY-MM-DD, 정확한 시점은 한국 시간 오프셋을 포함한 YYYY-MM-DDTHH:mm:ss+09:00을 사용합니다. 거래 시각처럼 날짜와 분리된 선택 시각은 HH:mm:ss 또는 null을 사용합니다."
  },
  "paths": {
...
    "/api/v1/monthly-budget/{month}": {
      "get": {
        "tags": [
          "월별 예산"
        ],
        "summary": "월별 예산 조회",
        "operationId": "get_api_v1_monthly_budget_month",
        "description": "Authorization: Bearer <JWT>로 인증한 사용자의 지정 연월 예산을 조회합니다. 저장된 예산이 없으면 amount:null을 반환합니다. query와 요청 본문은 허용하지 않습니다.",
        "parameters": [
          {
            "name": "month",
            "in": "path",
            "required": true,
            "description": "조회하거나 저장할 연월입니다. YYYY-MM 형식입니다.",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{4}-(?:0[1-9]|1[0-2])$",
              "example": "2026-09"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "정상 처리됨.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 200
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "month",
                        "amount"
                      ],
                      "properties": {
                        "month": {
                          "type": "string",
                          "pattern": "^[0-9]{4}-(?:0[1-9]|1[0-2])$",
                          "description": "예산을 적용할 한국 시간 기준 연월입니다. YYYY-MM 형식입니다.",
                          "example": "2026-09"
                        },
                        "amount": {
                          "type": [
                            "integer",
                            "null"
                          ],
                          "minimum": 0,
                          "maximum": 10000000,
                          "description": "해당 월의 예산입니다. 원화 정수 0~10,000,000원이며 미설정이면 null입니다.",
                          "example": 500000
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 200,
                  "code": "SUCCESS",
                  "message": "The request has been accepted and processed.",
                  "data": {
                    "month": "2026-09",
                    "amount": 500000
                  }
                }
              }
            }
          },
          "400": {
            "description": "month 형식이 유효하지 않음.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 400
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 400,
                  "code": "INVALID_REQUEST",
                  "message": "The request parameters or format are invalid.",
                  "data": null
                }
              }
            }
          },
          "401": {
            "description": "토큰이 제공되지 않았거나 검증에 실패함.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              },
              "WWW-Authenticate": {
                "schema": {
                  "type": "string",
                  "const": "Bearer"
                },
                "description": "Bearer 인증 안내"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 401
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 401,
                  "code": "AUTHENTICATION_REQUIRED",
                  "message": "Authentication is required or the token is invalid.",
                  "data": null
                }
              }
            }
          },
          "503": {
            "description": "DB 잠금 또는 일시 사용 불가.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 503
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 503,
                  "code": "SERVICE_UNAVAILABLE",
                  "message": "The service is temporarily unavailable.",
                  "data": null
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "월별 예산"
        ],
        "summary": "월별 예산 저장",
        "operationId": "put_api_v1_monthly_budget_month",
        "description": "Authorization: Bearer <JWT>로 인증한 사용자의 지정 연월 예산을 생성하거나 교체합니다. amount:null을 전달하면 해당 월의 예산을 미설정 상태로 바꿉니다. 성공하면 저장 결과를 반환합니다.",
        "parameters": [
          {
            "name": "month",
            "in": "path",
            "required": true,
            "description": "조회하거나 저장할 연월입니다. YYYY-MM 형식입니다.",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{4}-(?:0[1-9]|1[0-2])$",
              "example": "2026-09"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "월별 예산을 application/json 본문으로 전달합니다.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": [
                      "integer",
                      "null"
                    ],
                    "minimum": 0,
                    "maximum": 10000000,
                    "description": "해당 월의 예산입니다. 원화 정수 0~10,000,000원이며 미설정이면 null입니다.",
                    "example": 500000
                  }
                },
                "additionalProperties": false
              },
              "example": {
                "amount": 500000
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "정상 처리됨.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 200
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "month",
                        "amount"
                      ],
                      "properties": {
                        "month": {
                          "type": "string",
                          "pattern": "^[0-9]{4}-(?:0[1-9]|1[0-2])$",
                          "description": "예산을 적용할 한국 시간 기준 연월입니다. YYYY-MM 형식입니다.",
                          "example": "2026-09"
                        },
                        "amount": {
                          "type": [
                            "integer",
                            "null"
                          ],
                          "minimum": 0,
                          "maximum": 10000000,
                          "description": "해당 월의 예산입니다. 원화 정수 0~10,000,000원이며 미설정이면 null입니다.",
                          "example": 500000
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 200,
                  "code": "SUCCESS",
                  "message": "The request has been accepted and processed.",
                  "data": {
                    "month": "2026-09",
                    "amount": 500000
                  }
                }
              }
            }
          },
          "400": {
            "description": "month 또는 amount 형식이 유효하지 않음.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 400
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 400,
                  "code": "INVALID_REQUEST",
                  "message": "The request parameters or format are invalid.",
                  "data": null
                }
              }
            }
          },
          "401": {
            "description": "토큰이 제공되지 않았거나 검증에 실패함.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              },
              "WWW-Authenticate": {
                "schema": {
                  "type": "string",
                  "const": "Bearer"
                },
                "description": "Bearer 인증 안내"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 401
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 401,
                  "code": "AUTHENTICATION_REQUIRED",
                  "message": "Authentication is required or the token is invalid.",
                  "data": null
                }
              }
            }
          },
          "503": {
            "description": "DB 잠금 또는 일시 사용 불가.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 503
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 503,
                  "code": "SERVICE_UNAVAILABLE",
                  "message": "The service is temporarily unavailable.",
                  "data": null
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authorization: Bearer <JWT> 헤더로 인증합니다."
      }
    },
...
  }
}